+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,15 @@
 # docs
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @21stgov/commons-core@1.0.0
+  - @21stgov/commons-fonts@0.4.0
+  - @21stgov/commons-react@0.4.0
+  - @21stgov/commons-tokens@0.4.0
+
 ## 0.0.3
 
 ### Patch Changes